requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.0
drf-yasg==1.21.10
inflection==0.5.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
